[router] Helix-assisted scatter-gathering support in Venice Router
This code change introduced a new style of scatter-gathering for multi-key requests.
The idea behind this change is to limit the fanout size when scaling large-fanout cluster (venice-9 specifically).
With Helix group/zone feature(Helix Instance domain config), we will be able to keep a full replication
inside one zone/group if the total number of zone/group is same as the replication factor, which means
we could limit the fanout size to one zone/group no matter how big the storage node cluster could be.

Some details of the new scatter-gathering mode:
1. Router will assign an unique id to each Router request(the retry requests/scattered requests belonging
   to the same Router request will share the same request id).
2. {@link HelixAssistedScatterGatherMode} will assign a group id to each request in a round-robin fashion to guarantee
   the evenness across different groups.
3. If there is no healthy replica in the assigned group, it will try to find a healthy replica from the nearest group
   in one direction. For example, if there are 3 groups: 0, 1, 2 and group 1 is assigned to current request, and if
   there is no healthy replica in group 1, it will first look at group 2, then group 0.

This code change introduced several new configs to Router:
  /**
   * The Helix virtual group field name in domain, and the allowed values: "group" and "zone".
   * default: "group"
   */
  public static final String ROUTER_HELIX_VIRTUAL_GROUP_FIELD_IN_DOMAIN = "router.helix.virtual.group.field.in.domain";

  /**
   * This config is used to define the routing strategy for multi-key requests.
   * Please check {@text VeniceMultiKeyRoutingStrategy} to find available routing strategy.
   * Available options:
   * public enum VeniceMultiKeyRoutingStrategy {
   * // This mode will send the request to the first host of all the available replicas
   * GROUP_BY_PRIMARY_HOST_ROUTING,
   * // This mode will try send minimum of requests to storage node.
   * GREEDY_ROUTING,
   * // This mode will try to send the same key to the same partition replica.
   * KEY_BASED_STICKY_ROUTING,
   * // This mode will try to limit the fanout inside one helix group/zone.
   * HELIX_ASSISTED_ROUTING
   * }
   * Default: KEY_BASED_STICKY_ROUTING
   */
  public static final String ROUTER_MULTI_KEY_ROUTING_STRATEGY = "router.multi.key.routing.strategy";

This code change removed the following configs since they were not designed well to be mutually exclusive:
router.enable.sticky.routing.for.multi.get
router.greedy.multiget
But in cfg2, we are going to keep all the old/new configs to make it compatible with old/new releases for some time.

RB=2230814
BUG=VENG-6366
R=nithakka,xyin,shuhliu,zfan,anaberez,sdwu,xnma,jlliu,smaji,fvillega,mnhuang,zpolicze,yurzhou
A=sdwu