[cherry pick][server] hot fix for L/F write-compute stores
This is cherry pick hot fix commit on top of release-0.2.299-cutoff.
This commit include 2 fixes.
1. Leader SN chunking adapter mistakenly seralized the large value with
2 trailing bytes ignored.

2. Leader SN failed to kick off the latch and set all Replica into error
state.

This is an unusual commit that's intended to unblock our partners in
Prod. We will create a ne remote branch and it won't be committed to
remote master.

RB=2236782
G=voldemort-reviewers
R=shuhliu,mnhuang,goliu,fvillega
A=shuhliu,mnhuang