[server] Improve logging when rolling back a state transition
Currently when server state models roll back a state transition message,
it would log the exact same state transition again, and later log
that this state transition is completed, but actually it is the rollback
action that is completed instead of the original state transition message.

This RB improves the logging by explicitly mention the current action
is to roll back a state transition.

RB=2237042
BUG=VENG-6396
G=voldemort-reviewers
R=smaji,xyin,zpolicze,fvillega,zfan,jlliu,shuhliu,nithakka,goliu,sdwu,xnma,yurzhou,anaberez
A=anaberez