[router] Fixed a long tail retry bug with Helix Assisted Routing
The original logic will mark the passed storage node in the invocation
of `VenicePath#canRequestStorageNode`, which is error-prone.
`HelixAssistedScatterGatherMode` is invoking this function for every
replica, then make a decision according group distance, which means
the original request will put every replica as slow storage node, but
actually it will only send the request to a subset of all the storage nodes.

The fix is to explicitly add to the slow storage node set in `VeniceDispatcher`,
and `VenicePath#canRequestStorageNode` will be a read-only operation.

RB=2251667

RB=2251676
BUG=VENG-6439
R=nithakka,smaji,mnhuang,anaberez,yurzhou,zfan,sdwu,shuhliu,xnma,fvillega,zpolicze,jlliu,xyin
A=fvillega