[server] Add LeaderFollowerStateModelNotifier into notifiers queue
Latch was added between Offline to Follower transistion for current version.
However, it is not released in time because LeaderFollowerStateModelNotifier
is not added into notifiers queue.

The rb separates the notifier queue into two. One is for L/F and the other
is for O/O. Use the correct queue to create ingestion task.
LeaderFollowerStateModelNotifier is added to L/F queue. Common notifiers
like logNotifier are added to both queues.

RB=2252885
BUG=VENG-6426
G=voldemort-reviewers
R=zpolicze,jlliu,anaberez,xyin,fvillega,smaji,yurzhou,zfan,xnma,sdwu,goliu,nithakka,mnhuang
A=sdwu,mnhuang