[thin-client] GC optimization
Thin-client logic is relatively simple compared with other components
in Venice; thin-client gets some keys from users, serialize them and
send it out to backend. Most of the objects are created during serialization
for requests and deserialization for response; so the high level idea
for GC optimization is to reuse as much objects as we can during
ser/deserialization phase.

Introduced 2 new client configs:
1. reuseObjectsForSerialization
When set to true, the same BinaryEncoder and ByteArrayOutputStream instance
will be used for all kinds of serialization works, so that we don't need
to build a new oject for each request; noticed that each ByteArrayOutputStream
has its own byte array, which is one of the top big size items according
to heap dump.

The reused objects are thread local objects, so there is no risk for confliction

default value: false

2. useBlackHoleDeserializer
When set to true, the inner client type will become "AvroBlackHoleResponseStoreClientImpl"
which overrides the implementation for compute streaming, and drop
all the data chunk returned from Venice Routers. Basically it avoids
all kinds of deserialization works.

default value: false (this should be turned on for experiments only)

Other changes:
 - Introduced an array based primitive integer Set; it is optimized for
   GC by maintaining an int array without any boxing. Users of this set
   can reuse the same set if they want; "clear()" function of this set
   will keep the backing array for next time.
 - Enable fast-avro for serializing client requests if "useFastAvro"
   is set to true.

TODO:
The above 2 new client configs only works for compute streaming, we should
extend the support for single-get, batch-get and regular read compute.

RB=2254275
BUG=VENG-6455
G=voldemort-reviewers
R=yurzhou,zfan,shuhliu,xnma,goliu,anaberez,nithakka,zpolicze,fvillega,sdwu,smaji,xyin,jlliu
A=zpolicze