[server][da-vinci] Fixed a bug in ChunkedValueInputStream
The bug made it so that if -1 was present in the Avro serialized bytes,
it would cause the BinaryDecoder to enter a bad state and then fail to
deserialize properly.

The fix is to offset negative data by 256.

There is also an implementation of another function of the InputStream
class, the 3-args read() function which is a bit more performant.

RB=2254509
BUG=VENG-6456
G=voldemort-reviewers
R=nithakka,mnhuang,xyin,sdwu,yurzhou,jlliu,goliu,xnma,anaberez,zpolicze,shuhliu,zfan,smaji
A=shuhliu,goliu,zpolicze