[router] Fix nonsensical quota calculation when router count read doesn't make sense
This is to fix a known issue where reported quota drops to zero when zk connectivity is flakey

We do this in two ways.  One way defensively in the code which does the calculation.  Secondly, by making sure the number of routers reported by the router counter class is only based on what can be queryied by zk.

RB=2255771
BUG=VENG-6403
G=voldemort-reviewers
R=goliu,nithakka,xyin,fvillega,anaberez,zfan,jlliu,xnma,yurzhou,mnhuang,shuhliu,sdwu,smaji
A=mnhuang