[server] Misaligned internal ingestion states after Helix StateModel.reset()
Today the ingestion task and its internal components such as the consumer is reused when Helix StateModel.reset()
is called. reset() is invoked when the partition is resetting from ERROR state or when the SN is restarted or reconnected
to ZK. After reset the StateModel will go back to its initial state and start transitioning to top state again.
In the case when the SN is reconnected to ZK the following scenario can occur for low traffic RT stores with low lag requirements.

1. An ongoing ingestion task for a RT store's partition has consumed to end of VT at offset 200. Lag requirement is 100 and
the last synced (written to disk) offset is 50.

2. Some ZK connectivity issue occured and Helix resets the partition. The ingestion task is reused and subscribePartition(...)
is called as part of OFFLINE->BOOTSTRAP. As part of handling SUBSCRIBE we sync the in-memory partition state to the disk state
(offfset 50) but the subscribe action for the consumer is ignored because we already subscribed.

3. Now the partition consumption state is at offset 50 but consumer is at end of VT of offset 200. The replica cannot become ONLINE
because it does not satisfy the lag requirement and without new messages in the VT the ingestion task is not going to fetch any new
messages. The ingestion for this partition is now stuck in BOOTSTRAP.

RB=2257420
BUG=VENG-6397
G=voldemort-reviewers
R=xnma,fvillega,shuhliu,yurzhou,zfan,nithakka,smaji,sdwu,zpolicze,anaberez,jlliu,mnhuang,goliu
A=mnhuang