[controller][server] Add producer time lag threshold as a hybrid config
Currently hybrid lag computation formula for Online/Offline mode is:
lag = (RTMaxOffset - RTReplayOffset) - (VTEndOffset - VTStartOfBufferReplayOffset)
This works based on 2 assumptions:
1. Messages in RT from RTReplayOffset to RTMaxOffset would not be deleted;
2. If there are X number of messages from RTReplayOffset to RTMaxOffset, X number of
   messages will be replicated to VT.

Assumption #1 is wrong if Kafka log compaction is enabled for RT; assumption #2 is
wrong if Brooklin replicates duplicated messages, which cannot be controlled by Venice.

To solve these issues, this RB adds a third hybrid store config which is the producer
time lag threshold. Servers will not report ready-to-serve for a specific partition
until it sees a message whose producer timestamp if fresh enough that the difference
between current time and the producer timestamp is smaller than the time lag threshold.

By default, the time lag threshold is -1, meaning servers would only check offset lag
for reporting ready-to-serve.

If offset lag threshold is -1, servers will only check time lag for report online.

This new hybrid config is not needed for L/F model because when calculating hybrid lag
in L/F model, the lag is the (RTMaxOffset - LeaderOffsetInRT), which is accurate.
However, this can serves as an additional feature if offset lag doesn't make sense to
users; so we can ask users to input an acceptable time lag instead of offset lag while
creating hybrid stores.

Note for SRE:
Noticed that admin operation protocol is also upgraded to v45 in this commit, so child
controllers must be deployed before parent controllers.

Other changes:
1. Remove the hotfix which made time lag threshold as a server config;
2. Extended test AdminOperationProtocolCompatibilityTest to make sure the admin protocol
   evolution also works for fast avro.

RB=2283813
BUG=VENG-6500
G=voldemort-reviewers
R=xhao,jlliu,anaberez,smaji,xyin,goliu,ssen,sdwu,nithakka,xnma,zpolicze,shuhliu,fvillega,yurzhou
A=smaji,xyin