[common][controller] Enabling KafkaAdminClient in controller instead of ScalaAdminUtils for all kafka related operations
1. Simplified Kafka cleanup service since we no longer need to insepect Kafka Zk internal states to confirm
the deletion of a topic when using Java kafka admin client. In addition confirmed with the Kafka team that
previous concerns around topic deletion on nonexistent topic and topics in the process of getting deleted are
now handled gracefully.
  - UnknownTopicOrPartitionException will be thrown when try to get the future in the delete result of a nonexistent topic
  - No error when deleting a topic that's in the process of getting deleted but the future will complete only after it's
    deleted successfully.

2. High risk change due to many operations now return a future and previously known Kafka exceptions are now wrapped in
ExecutionException when getting the future/result. This means we'd have to manually extract these exceptions or handle them
differently in various locations when we interact with Kafka through the new Java KafkaAdminClient.

RB=2282816
BUG=VENG-6391
G=voldemort-reviewers
R=xhao,sdwu,smaji,xnma,jlliu,mnhuang,shuhliu,zpolicze,nithakka,yurzhou,ssen,anaberez,goliu,fvillega
A=sdwu