[common] Refresh the store if version does not exist when new offline push comes in
Sometimes L/F store router could not switch to the new version because it detected uncompleted partition.
Actually all partitions were completed. It was because the router did not subscribe to partition status
change. From heapdump, topicToResourceStatus map contained the topic, which indicates handleChildChange
was invoked. We use version L/F config to determine whether to subscribe. It is possible that the offline
push update came earlier than the store version update. Store version did not exist yet when new offline
push arrived.

To handle this race condition: when new offline push arrives, if the version does not exist, refresh the
store. The refresh logic is similar to that in KafkaStoreIngestionService#createConsumerTask.

RB=2293735
BUG=VENG-6253
G=voldemort-reviewers
R=jlliu,smaji,xyin,anaberez,zpolicze,goliu,sdwu,nithakka,xhao,yurzhou,ssen,xnma,mnhuang,fvillega
A=sdwu