[router] Fix NPE while reading partition assignment
When there is no online replicas for a specific partition, VeniceVersionFinder
wants to include the current partition assignments in its error message.
If the router hasn't received the callback for the partition assignment,
it's a valid reason that there is no online replicas; however, in this case,
NPE is thrown while reading partition assignment.

This RB fixes the issue by returning EMPTY map if there is no assignment;
if the expectation about function "HelixBaseRoutingRepository#getAllInstances"
is not only for logging purpose anymore. We need to refresh the partition
assignment for the store.

RB=2296953
BUG=VENG-6492
G=voldemort-reviewers
R=shuhliu,yurzhou,ssen,xnma,xhao,xyin,zpolicze,anaberez,jlliu,fvillega,sdwu,smaji,goliu,nithakka
A=yurzhou,sdwu