[server] Fixed one RocksObject leak
This code change fixed one RocksObject leak: WriteOptions, which is per partition.
And it also adopted the new byte array api in the read path to improve GC.

In the write path, there is an api to accept byte array with offset/length, but
it is only available in RocksDB.put, but not SSTFileWriter.put.
I am planning to use another RB to adopt it since it is a little complicated.

RB=2297629
BUG=VENG-6544
R=xhao,fvillega,yurzhou,sdwu,zpolicze,ssen,xyin,mnhuang,nithakka,anaberez,shuhliu,smaji,xnma,jlliu
A=fvillega