[server] Ingestion optimization for L/F model
Currently there is an ingestion bottleneck in function
"checkLongRunningTaskState" which is invoked in each consumption
iteration. The function would call storeBufferService.drainBufferedRecordsFromTopicPartition
in order to get the latest consumption timestamp for a partition;
as a result, L/F model would flush messages into RocksDB after
consuming a batch of records from Kafka, which reduces throughput
dramatically.

This RB removes the call to storeBufferService.drainBufferedRecordsFromTopicPartition
because we don't need to wait for the drainer service to update
the latest consumption task. We can update the consumption task
after consuming the records.

Other changes:
1. Instead of using "partition.lastUpdate" to record the latest consumption
   time, use an in-memory state "latestMessageConsumptionTimestampInMs"
   since we don't rely on this time to do any recovery;
2. The "partition.lastUpdate" metadata is used to record the latest
   producer timestamp for each partition now, since producer time
   lag can be used to check whether a partition can go online, we
   should persist the information to disk.

RB=2303509
BUG=VENG-6549
G=voldemort-reviewers
R=ssen,nithakka,anaberez,yurzhou,jlliu,sdwu,shuhliu,smaji,xnma,xhao,xyin,zpolicze,goliu,fvillega
A=goliu