[da-vinci] Bug fix: avoid duplicate partition subscription
Currently StoreIngestionTask is not able to subscribe to an already
subscribed partition twice, because the in-memory DIV state will
be override while consumer refuses to subscribe to a lower offset.

This RB fixes the issue by adding a safe guard in Da Vinci, if the
VersionBackend already contains a future for a partition, it means
the partition has been subscribed already, so VersionBackend#subscribeSubPartition
will return the future without subscribing again.

TODO: Fix the core issue in StoreIngestionTask where the same partition
can not be subscribed twice: VENG-6636

RB=2321402
BUG=VENG-6635
G=voldemort-reviewers
R=xyin,xhao,zpolicze,xnma,goliu,sdwu,nithakka,shuhliu,ssen,smaji,yurzhou,anaberez,fvillega,jlliu
A=anaberez