[controller][router] fix partition status sorted by enum
Recently we found that in queries, router doesn't have available instances even if
the replicas are all COMPLETED for L/F stores.

This was because the available replicas used to have ERROR in the history status.
So when the replica has ERROR -> STARTED -> COMPLETED in the history statuses in offlinepush,
the status ERROR will be returned because ERROR is "larger" than "COMPLETED" in enum.

This rb fixed it by directly reversing the historic status list so it will read
from the latest status.

RB=2322421
BUG=VENG-6573
G=voldemort-reviewers
R=nithakka,xnma,mnhuang,goliu,fvillega,shuhliu,zpolicze,xyin,jlliu,ssen,smaji,xhao,anaberez,sdwu
A=shuhliu