[controller] Bug fix: NPE without detailed message is mistakenly ignored in parent controller
When an update_store command is sent to parent controllers for a store
that doesn't exist, NullPointerException would happen while getting
the store owner, but the NPE doesn't have any detailed message, so when
we set error message in ControllerResponse, nothing is specified and
thus AdminTool thinks the command succeeds.

This RB fixes the bug by explicitly throwing a VeniceException when
the store doesn't exist, instead of waiting for a NPE; also, specify
customized error message for VeniceResponse rather than filling it
with exception message only.

RB=2324823
BUG=VENG-6378
G=voldemort-reviewers
R=goliu,jlliu,sdwu,xyin,xnma,zpolicze,fvillega,shuhliu,nithakka,ssen,yurzhou,anaberez,smaji,xhao
A=nithakka