[da-vinci] Only main ingestion task should complete subscribe future
This is to avoid the deadlock situation when the user calls unsubscribe while the main ingestion
task is still processing the subscribe message from the consumer actions queue. Both operations
have synchronized blocks in VersionBackend. The unsubscribe method could take the lock and wait
for consumption to stop but it will block the subscribe action from completion and create a deadlock
because unsubscribe consumer action can only be processed after the subscribe action before it.

RB=2327745
G=voldemort-reviewers
R=xnma,jlliu,fvillega,sdwu,yurzhou,smaji,zpolicze,ssen,nithakka,shuhliu,goliu,anaberez,xhao,mnhuang
A=jlliu