[server] Fix follower_offset_lag stats and add new stats "write_compute_operation_failure"
We recently added a new stat "follower_offset_lag". However, the stat is
not available in the dashboard due to a NPE issue. This is because the
versioned stats framework hooks stats reporters and the stats
dynamically. The previous reporter tried to register some optional stats based
on stat's conditions. But these stats have not been hooked with the
reporter yet.

Along with the fix. The RB also adds a new stat to monitor write_compute
failure.

RB=2327815
R=xyin,zpolicze,smaji,fvillega,goliu,ssen,mnhuang,xhao,shuhliu,xnma,jlliu,anaberez,nithakka,yurzhou
A=yurzhou