[common] PartitionStatusOnlineInstanceFinder stops refreshing store if version is deleted
Sometimes TestBatchForRocksDB.testDuplicateKey cannot finish in 60 seconds. Since at first
duplicate keys are not allowed, the version becomes error version and gets deleted quickly.
If version is already deleted by the time PartitionStatusOnlineInstanceFinder handles new
push status, no matter how many times the store is refreshed, version is still not present.
As a result, router wastes time on refreshing store and synchronized methods are blocked.

This rb ends store refresh earlier when it detects that version does not exist but version
number is less than or equal to the largest used version number, which indicates that the
version is deleted.

RB=2330528
BUG=VENG-6571
G=voldemort-reviewers
R=xyin,zpolicze,goliu,yurzhou,fvillega,smaji,xnma,jlliu,mnhuang,anaberez,xhao,nithakka,ssen,sdwu
A=anaberez,sdwu