[server] If remote consumption flag is false, messages in the drainer queue should be processed before that
Currently there is an edge case in native replication:
1. Ingestion thread would turn off remote consumption flag when EOP is received
2. Drainer thread is responsible for processing messages and updating offset
So there is a chance that remote consumption flag is turned off right
after EOP is processed but before End_of_Segment (EOS) is processed;
as a result, drainer thread chooses to update the offset metadata with
the EOS offset from remote topic and not to produce this message to
local version topic.

This RB fixes the edge case with two changes:
1. if remote consumption flag is changed from true to false, messages
   for the partitions in the drainer queue should be processed before that.
2. As long as the Start_of_Segment and End_of_Segment messages are not
   from real-time topic, we will try to do an extra check and potentially
   produce that message to local version topic (an extra check will avoid
   producing a message from local VT back to local VT)

Also extended the current integration test to make sure when native replicaiton
completes, offset in metadata service will always be smaller than or equal to
the endOffset in the local version topic.

RB=2348823
BUG=VENG-6710
G=voldemort-reviewers
R=goliu,xhao,fvillega,shuhliu,smaji,anaberez,ssen,xnma,yurzhou,sdwu,zpolicze,nithakka,xyin,jlliu
A=ssen