[fast-client] Routing/Fault tolerance support in Fast-client
This code change includes the routing/fault tolerance support.
Fault tolerance:
This code change is using the similar way as Venice Router to leverage
query depth(pending request count) to decide the healthiness of each replica.
For the good response, the pending request counter will be reset as long as
the response is received.
For the error response, a delay is specified when resetting the pending request
counter, so that the query depth of the unhealthy replicas will be higher.
When the query depth excceeds the defined threhold, the replica will be marked
as blocked.
So far, this code change doesn't implement any heart-beat logic to proactively
monitor the healthiess of each replica since it is too expensive, but to leverage
the users' requests to detect the degraded instances.

Routing:
This code change only implements the least-loaded routing strategy.
1. If the weight of every instance is same, it will fall back to the round-robin fashion.
2. This strategy will skip any blocked instance.
3. When any selected instance is marked as unhealthy, this strategy will try to back-fill with the healthy instances,
   and there are two purposes:
   a. The latency shouldn't be affected since this strategy will still try to send request to the required healthy instances.
   b. The unhealthy instance will still receive any requests, so we could mark it healthy once it is recovered.

This code change also includes the per-instance metrics.
These metrics will be very helpful in end-to-end testing to monitor the behavior of routing and fault tolerance.
If we think the per-instance metrics are too many, we could disable/remove them in the future.

RB=2358256
BUG=VENG-6653
R=sdwu,xyin,shuhliu,smaji,yurzhou,anaberez,ssen,nithakka,xnma,jlliu,zpolicze,fvillega,mnhuang,xhao
A=smaji