[server] Disabled byte buffer reuse for SST File writer
With L/F model, the total number of RocksDB writer threads is not limited
since the RocksDB update is triggered in Kafka producer callback thread,
so the total amount of direct memory required for byte buffer reuse
with SSTFileWriter is not limited either since the direct memory
pre-allocation is per writer thread.

Once the L/F model refactoring gets committed, we could re-enable
this byte buffer optimization since the drainer threads are the
only writers.

RB=2365936
BUG=VENG-6751
R=fvillega,mnhuang,zpolicze,jlliu,xnma,shuhliu,yurzhou,ssen,nithakka,anaberez,smaji,xhao,xyin,sdwu
A=fvillega,zpolicze,xyin