Controller deadlock when transition from LEADER->STANDBY due to lock contention with push monitor
Changed the locking behavior in onExternalViewChange in AbstractPushMonitor. No longer taking the pushMonitor
lock at the very beginning. This get rid off the nested locking where we held the pushMonitor lock while waiting
for another lock. The calls to update the status updatePushStatus will take and release the pushMonitor lock so
the in memory data structure topicToPushMap is still protected.

The risk is that now multiple onExternalViewChange event can theoretically be processed in parallel. The intermediate
steps will still be sequential due to locking in other components. EV events are currently triggered by a single thread,
as long as we keep it this way the risk is acceptable.

RB=2366053