[server][test] Notify the waiter inside SharedKafkaConsumer#unSubscribe if the consumer does not have subscription
TestHybrid.testHybridStoreQuota is flaky when leader/follower and shared consumer are enabled. Sometimes it fails
because the consumer does not poll anymore after unsubscribing from VT during follower to leader transition. In
theory, the consumer should subscribe to VT again. However, SharedKafkaConsumer#unSubscribe is hanging and waiting
for a invocation of SharedKafkaConsumer#poll. But ConsumptionTask#run does not poll anymore since the consumer
does not have any subscription.

The fix is to invoke SharedKafkaConsumer#poll no matter whether the consumer has subscription or not. If it does
not have subscription, sleep and return.

RB=2367115
G=voldemort-reviewers
R=mnhuang,yurzhou,zpolicze,sdwu,fvillega,ssen,goliu,xhao,xyin,jlliu,nithakka,xnma,anaberez,smaji
A=goliu,mnhuang,ssen