[common][router][server] Store-level access control support in storage node
This code change adds the store-level access control support in storage node tier.
It means if this store-level access controller is enabled, fast-client could
access storge node directly, and with two access controllers, it can be run
in 3 different modes in prod.
1. Only {@link ServerAclHandler} to only allow requests from Routers (static ACL for Routers).
2. Only {@link ServerStoreAclHandler} to only allow reqeusts from Application/Client (dynamic store ACL).
3. Both {@link ServerAclHandler} and {@link ServerStoreAclHandler} to allow requests
   from both Routers and Applcations/Client directly (static ACL for Routers + dynamic store ACL).

When starting testing fast-client, we need Server run in 3rd mode.

RB=2368099
BUG=VENG-6754
R=jlliu,mnhuang,xyin,yurzhou,xhao,zpolicze,shuhliu,sdwu,xnma,ssen,anaberez,nithakka,fvillega,smaji
A=zpolicze