[da-vinci][server] Live updates suppression support
This RB adds support for live updates suppression in da-vinci. Once
the feature is turned on, consumption will stop for a partition once
it's ready to serve, or consumption will not start if local data
exists for the subscribed partitions.

To turn on the feature, users can set the below backendConfig to true:
freeze.ingestion.if.ready.to.serve.or.local.data.exists

Or enable it through DaVinciConfig programatically:
DaVinciConfig#freezeIngestionIfReadyToServeOrLocalDataExists(true)

This is an opt-in feature.

RB=2376682
BUG=VENG-6696
G=voldemort-reviewers
R=jlliu,ssen,fvillega,sdwu,smaji,zpolicze,xnma,xyin,shuhliu,anaberez,nithakka,yurzhou,goliu,xhao
A=anaberez,smaji