[common][controller] Revert the in place schema change of metadadata system store
Since the current value schema of metadata system store is already being used
in prod, we shouldn't change this schema in place since it will cause
de-serialization failure if the message is writen with the updated schema with
the same schema id.
To be safe, every schema change should be done via proper schema evolution, which
means we should add a new version of value schema for every change.

RB=2375733
BUG=VENG-6795
R=jlliu,zpolicze,smaji,xhao,mnhuang,anaberez,shuhliu,ssen,nithakka,fvillega,xyin,xnma,yurzhou,sdwu
A=mnhuang