[client][common] Defensive check to stop producing chunked PUT in RT Topic
Following are the current constraint with respect to producing chunked
messages. We do not support producing chunked message in RT topics.
Chunked message is only supported for PUT in VT topic. It also means
samza streaming must not produce large enough PUT which may need to be chunked.
For incremental push if it's producing to VT it is allowed to produce large PUT value
that may get chunked
but if incremental push it's producing to RT (for write-compute
enabled stores) then again it must not produce large PUT value which
need to be chunked.
Also for reprocessing job it should not produce large PUT which needs to
be chunked.

This change add a simple defensive check along with already present
checks in VeniceWriter to enforce all of the above.

RB=2373237
BUG=veng-6702
G=voldemort-reviewers
R=xnma,goliu,sdwu,mnhuang,shuhliu,jlliu,xyin,yurzhou,zpolicze,fvillega,nithakka,smaji,xhao,anaberez
A=mnhuang