[common] Buffer replay should be allowed to start even if PROGRESS is reported after END_OF_PUSH_RECEIVED
This is an attempt to fix the edge case encountered in production where all replicas of a partition is in
PROGRESS state after END_OF_PUSH_RECEIVED and buffer replay cannot be started since the old logic will require
at least one replica's current statatus is END_OF_PUSH_RECEIVED. That's a bad assumption regardless if we are
missing ZK events or not because if the controller restarted or a leadership transfer occurs while the SNs
continued to make progress and reach the steady state of all replicas in PROGRESS after END_OF_PUSH_RECEIVED
status then even bouncing the controller will not kick off the buffer replay when it really should.

This edge case is problematic for metadata system store auto creation because the edge case is more likely to
occur with frequent empty pushes with very few replicas.

RB=2384598
BUG=VENG-6723
G=voldemort-reviewers
R=goliu
A=anaberez,goliu