[common] Metadata system store auto-creation
This code change is leveraging `SystemSchemaInitializationRoutine` to
create zk shared metadata system store when Controller is up and running.
And this will only create zk shared metadata system store the system
store cluster, but not in every Venice cluster, and the reason is that
the key/value schemas can be shared across all the Venice clusters, and
there is no need to manage the versions for this zk shared metadata
system store.

RB=2385282
BUG=VENG-6781
G=voldemort-reviewers
R=xnma,xyin,fvillega
A=xyin