[H2V][Controller] Add a new param "data_center_name" to topic_request rest call
"topic_request" is a rest call that H2V uses to get the topic name it
will write data into from Controller. On the other side, Controller
will create new Kafka topic and the coressponding version. This change
adds a new param to identify where the push happens. This helps
Controller locate the best Kakfa cluster to use if native replication is
enabled.

RB=2391095
R=anaberez,mnhuang,jlliu,xyin,yurzhou,xnma,zpolicze,nithakka,smaji,shuhliu,ssen,fvillega,goliu,xhao
A=mnhuang