[Server] Release ingestion latch while resetting state model
While transiting from BOOTSTRAP to ONLINE in O/O model, or OFFLINE
to STANDBY for current version in L/F model, a latch is placed to
wait for the ingestion to complete; however, if ZK disconnection
and reconnection happens, Helix will invoke reset() API which
unsubscribes from the topic, but reset doesn't release the latch
which keeps hanging the original transition.

This RB adds support for releasing latch (if any) when unsubscribing
from the topic.

RB=2395799
BUG=VENG-4377
G=voldemort-reviewers
R=yurzhou,nithakka,ssen,xhao,goliu,xyin,smaji,jlliu,sdwu,fvillega,anaberez,shuhliu,zpolicze,xnma
A=sdwu