[samza][controller][server][h2v] Fixed infinite loop inside VeniceWriter
Currently there is an infinite loop inside VeniceWriter; when a segment
has lasted over 5 minutes, we call endSegment proactively in order to
create small segments; however, endSegment needs to get current segment,
which calls endSegment again.

This RB fixed the bug by setting the currentSegmentCreationTime to -1 to
indicate that this segment is ended, so no need to eng it again.

Besides, added an integration test to exercise the small segment creation
logic aggressively.

RB=2408826
BUG=VENG-6859
G=voldemort-reviewers
R=smaji,xnma,jlliu,xyin,sdwu,shuhliu,zpolicze,nithakka,yurzhou,goliu,fvillega,xhao,anaberez,ssen
A=yurzhou