[samza] Use running fabric of Samza Reprocessing job in A/A native replication source fabric
This change sets the native replication source fabric to the running fabric
where the instance of the Samza Reprocessing(SR) job is running. Even
though the change is generic, the value set to batchStartingFabric parameter of
requestTopicForWrites() will not be used in O/O mode or regular
hybrid streaming mode. It will only be used when SR job is running in
native replication mode.
Samza job passes this metadata config "com.linkedin.app.env" in Config object
and as well as property variable. Checking the value in both places if one happens
to be not set.
Also if the samza job happens to run on corp cluster we will force the
native replication source fabric to prod-lva1 as this is our default
choice.

RB=2410626
BUG=veng-6864
G=voldemort-reviewers
R=anaberez,zpolicze,goliu,jlliu,xhao,mnhuang,smaji,sdwu,shuhliu,xyin,yurzhou,xnma,fvillega,nithakka
A=mnhuang