[Common] Reduce the callback EV triggers
When HelixExternalViewRepository receives a routing table snapshot from
Helix, it used to trigger the callbacks for every resource even if the
resource remains the same in the new snapshot. This RB asks the
HelixExternalViewRepository to calculate the changes between current
Resource assignment and the new assignment. Thus, it reduces the number
of callbacks.

RB=2412558
R=smaji,mnhuang,jlliu,goliu,yurzhou,zpolicze,fvillega,xyin,shuhliu,ssen,xhao,nithakka,anaberez,xnma
A=yurzhou