[controller][server] Replica status reporting from Venice SN
This code change added support to report replica status from Venice SN.
So far, it is still using map field in `StoreMetaValue` to store replica
statuses since the nullable union write compute support is still problematic.
Once the nullable union write compute support is ready, we could switch
to it.
Right now, this code change didn't implement a way to clean up the replica
status for the deprecated versioins, and the replica status entries for
the deprecated version will contains empty map, and we are in the middle
of discussing whether we could introduce a new write compute feature
to clean it up automatically when the replica status map becomes empty.

For leadership handover, this code change adopts the fast handover since
the freshness of replica status system store is critical for read request
routing, and this is controlled by a new config in SN:
server.system.store.promotion.to.leader.replica.delay.seconds: default 1 second

TODO:
1. Store deletion support.
2. Store migration support.

RB=2424109
BUG=VENG-6880
G=voldemort-reviewers
R=sdwu,xyin
A=xyin