[server] fix error replicas triggered by hybrid store quota in rebalance
Per current implementation of hybrid storage quota - when the partition-level quota is reached,
we will pause partition ingestions. This is fine. Online replicas for current version wouldn't
become back to bootstrapping or error.

But if rebalance happens, then the new replica for online version could become error.
This is because the ingestion is paused and the new replica could not become completed.
So it will time out and then go to error state.

This rb fixes this by directly reporting online when we find a bootstraping replica with online version.

RB=2422540
BUG=VENG-6661
G=voldemort-reviewers
R=apoursam,anaberez,ssen,fvillega,xhao,shuhliu,mnhuang,smaji,zpolicze,sdwu,nithakka,goliu,xyin,jlliu,xnma
A=nithakka